Extract folder browser path helpers
